Add a simple example
