+++ examples/streaming_real/Cargo.toml
@@ -12,4 +12,5 @@ tokio-stream = "0.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 async-trait = "0.1"
-colored = "2.0" 
+colored = "2.0"
+clap = { version = "4.0", features = ["derive"] }
